LayoutTests/imported/w3c/web-platform-tests/offscreen-canvas/text/2d.text.font.default.worker-expected.txt


FAIL 2d assert_equals: ctx.font === '10px sans-serif' (got [undefined], expected 10px sans-serif[string]) expected (string) "10px sans-serif" but got (undefined) undefined
